requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
langchain-groq
langchain-openai
langchain-community
langchain-huggingface
langgraph
faiss-cpu
edge-tts
groq